fix: prevents setting LoadBalancerIP with AWS annotations
Addresses #688, this commit adds a check to the ControlPlane service's
annotations.
If the `spec.loadBalancerIP` property is set in the service, the AWS
cloud controller will complain and causes issues.

The controller now checks for the presence of these specific annotations
and avoids setting the LoadBalancerIP to ensure that the service is correctly
configured according to AWS requirements.